// src/app/models/PeriodicElement.ts
export class PeriodicElement {
  evalId: number;
  evalProvee: number;
  evalNit: string;
  evalDescripcion: number;
  evalCumplimiento1: number;
  evalCumplimiento2: number;
  evalCalidad1: number;
  evalCalidad2: number;
  evalAcompana1: number;
  evalAcompana2: number;
  evalAdmon1: number;
  evalAdmon2: number;
  evalAdmon3: number;
  evalAdmObserva: string;
  gaurecomendariaProveedor: string;
  gauNombreEvaluador: string;
  gauCargoEvaluador: string;
  userProviderFecha : Date;

}